maxHill for checkpoint, settings editting, indicator fixes, name input field script

+++ Assets/Scripts/ForGUI/MainMenu.cs
@@ -5,10 +5,19 @@ public class MainMenu : MonoBehaviour
 {
     public string startSceneName = "LevelCave000";
 
+    void Awake()
+    {
+        if (!PlayerPrefs.HasKey("Initialized"))
+        {
+            PlayerPrefs.DeleteAll();
+            PlayerPrefs.SetInt("Initialized", 1);
+            PlayerPrefs.Save();
+        }
+    }
+
     public void StartGame()
     {
         DataContainer.checkpointIndex = new Vector3(-55, 25, 0);
-        PlayerPrefs.DeleteAll();
         //for test
         PlayerPrefs.SetInt("coins", 200);
         // for test
